+++ src/disopy/cogs/queue.py
@@ -9,7 +9,7 @@ from collections import deque
 from typing import Iterable, NamedTuple, cast
 
 import discord
-from discord import VoiceClient, app_commands
+from discord import PCMVolumeTransformer, VoiceClient, app_commands
 from discord.ext.commands import Bot
 from discord.interactions import Interaction
 from knuckles import Subsonic
@@ -17,8 +17,6 @@ from knuckles import Subsonic
 from ..config import Config
 from ..options import Options
 from .base import Base
-
-logger = logging.getLogger(__name__)
 
 
 class Song(NamedTuple):
@@ -31,6 +29,9 @@ class Song(NamedTuple):
 
     id: str
     title: str
+
+
+logger = logging.getLogger(__name__)
 
 
 class Queue:
@@ -145,66 +146,36 @@ class QueueCog(Base):
         self.config = config
 
         self.queue = Queue()
+        self.now_playing: Song | None = None
 
         self.skip_next_autoplay = False
 
-    async def get_voice_client(self, interaction: Interaction) -> VoiceClient | None:
-        """Get the voice client of an interaction.
-
-        Args:
-            interaction: The interaction to get the voice client from.
-
-        Returns:
-            The voice client or None if the interaction doesn't have one attach to it.
-        """
-
-        if interaction.guild is None:
-            await self.send_embed(interaction, "We're not taking in a guild!")
-            return None
-
-        if interaction.guild.voice_client is None:
-            await self.send_embed(interaction, "I'm not connected to a channel!")
-            return None
-
-        return cast(VoiceClient, interaction.guild.voice_client)
-
-    async def is_not_in_channel(self, interaction: Interaction, connect: bool = False) -> bool:
-        """Check if the user is with the bot in the same voice channel.
-
-        Args:
-            interaction: The interaction to get the voice channel from.
-            connect: If the bot should automatically connect to the channel where the user is.
-
-        Returns:
-            If the user is with the bot in the same voice channel.
-        """
-
+    async def get_voice_client(self, interaction: Interaction, connect: bool = False) -> VoiceClient | None:
         user = interaction.user
         if isinstance(user, discord.User):
-            await interaction.response.send_message(
-                "You are not a member of the guild, something has gone very wrong..."
-            )
-            return True
+            await self.send_error(interaction, ["You are not a member of the guild, something has gone very wrong..."])
+            return None
 
         if user.voice is None or user.voice.channel is None:
-            await interaction.response.send_message("You are not connected to any voice channel!")
-            return True
+            await self.send_error(interaction, ["You are not connected to any voice channel!"])
+            return None
 
         guild = interaction.guild
         if guild is None:
-            await interaction.response.send_message("We are not chatting in a guild, something has gone very wrong...")
-            return True
+            await self.send_error(interaction, ["We are not chatting in a guild, something has gone very wrong..."])
+            return None
 
         if guild.voice_client is None:
             if connect:
-                await user.voice.channel.connect(self_deaf=True)
-            return False
+                return await user.voice.channel.connect(self_deaf=True)
+            await self.send_error(interaction, ["I'm not connected to a voice channel!"])
+            return None
 
         if user.voice.channel != guild.voice_client.channel:
-            await interaction.response.send_message("Join the same voice channel where I am")
-            return True
+            await self.send_error(interaction, ["Join the same voice channel where I am"])
+            return None
 
-        return False
+        return cast(VoiceClient, guild.voice_client)
 
     def play_next_callback(self, interaction: Interaction, exception: Exception | None) -> None:
         """Callback called when starting the playback of the next song in the queue.
@@ -257,14 +228,34 @@ class QueueCog(Base):
         voice_client = cast(VoiceClient, interaction.guild.voice_client)
 
         voice_client.play(
-            discord.FFmpegPCMAudio(str(song_path.absolute()), options=f'-filter:a "volume={self.config.volume / 100}"'),
+            discord.FFmpegPCMAudio(str(song_path.absolute())),
             # The ugliest Python syntax ever added to the language. The only way to do an conditional in a lambda
             # is with ternary operator, the value it returns is meaningless
             after=lambda exception: self.play_next_callback(interaction, exception),
         )
 
-    @app_commands.command(description="Play a song")
-    async def play(self, interaction: Interaction, song_name: str) -> None:
+        if voice_client.source is None:
+            logger.error("The source is not available to attach a volume transformer!")
+            return
+
+        self.now_playing = song
+        voice_client.source = PCMVolumeTransformer(voice_client.source, volume=self.config.volume / 100)
+
+    @app_commands.command(description="Add a song, album, or playlist to the queue")
+    @app_commands.choices(
+        what=[
+            app_commands.Choice(name="Song", value="song"),
+            app_commands.Choice(name="Album", value="album"),
+            app_commands.Choice(name="Playlist", value="playlist"),
+        ]
+    )
+    async def play(
+        self,
+        interaction: Interaction,
+        query: str,
+        # Ignore the MyPy error because discord.py uses the type to add autocompletion
+        what: app_commands.Choice[str] = "song",  # type: ignore
+    ) -> None:
         """Add a song in the queue and start the playback if it's stop.
 
         Args:
@@ -272,25 +263,79 @@ class QueueCog(Base):
             song_name: The name of the song.
         """
 
-        if await self.is_not_in_channel(interaction, True):
-            return
+        # Extract the type of element to be search, taking care of the default value
+        choice = what if isinstance(what, str) else what.value
 
-        songs = self.subsonic.searching.search(song_name, song_count=10, album_count=0, artist_count=0).songs
-        if songs is None:
-            await self.send_embed(interaction, f"No song found with the name: {song_name}")
-            return
-
-        song = songs[0]
-        if song.title is None:
-            await self.send_embed(interaction, f"The song is missing required metadata: {song_name}")
-            return
-
-        self.queue.append(interaction, Song(song.id, song.title))
-        await self.send_embed(interaction, "Feel the beets... 🎵", [f"Playing: **{song.title}**"])
-
-        voice_client = await self.get_voice_client(interaction)
+        voice_client = await self.get_voice_client(interaction, True)
         if voice_client is None:
             return
+
+        playing_element_name = query
+        first_play = self.queue.length(interaction) == 0 and self.now_playing is None
+
+        match choice:
+            case "song":
+                songs = self.subsonic.searching.search(query, song_count=10, album_count=0, artist_count=0).songs
+                if songs is None:
+                    await self.send_error(interaction, [f"No songs found with the name: **{query}**"])
+                    return
+
+                song = songs[0]
+                if song.title is None:
+                    await self.send_error(interaction, [f"The song is missing the required metadata: {query}"])
+                    return
+
+                playing_element_name = song.title
+                self.queue.append(interaction, Song(song.id, song.title))
+
+            case "album":
+                albums = self.subsonic.searching.search(query, song_count=0, album_count=10, artist_count=0).albums
+                if albums is None:
+                    await self.send_error(interaction, [f"No albums found with the name: **{query}**"])
+                    return
+
+                album = albums[0].generate()
+                print(album)
+                if album.songs is None:
+                    await self.send_error(interaction, [f"The album is missing the required metadata: {query}"])
+                    return
+
+                if album.name is not None:
+                    playing_element_name = album.name
+
+                for song in album.songs:
+                    if song.title is None:
+                        logger.error(f"The song with ID '{song.id}' is missing the name metadata entry")
+                        continue
+
+                    self.queue.append(interaction, Song(song.id, song.title))
+
+            case "playlist":
+                for playlist in self.subsonic.playlists.get_playlists():
+                    if playlist.name is None:
+                        continue
+
+                    if query in playlist.name:
+                        playlist = playlist.generate()
+                        if playlist.songs is None:
+                            await self.send_error(interaction, ["The playlist has no songs!"])
+                            return
+
+                        if playlist.name is not None:
+                            playing_element_name = playlist.name
+
+                        for song in playlist.songs:
+                            if song.title is None:
+                                logger.error(f"The song with ID '{song.id}' is missing the name metadata entry")
+                                continue
+
+                            self.queue.append(interaction, Song(song.id, song.title))
+                        break
+
+        if first_play:
+            await self.send_answer(interaction, "🎵 Now playing!", [f"**{playing_element_name}**"])
+        else:
+            await self.send_answer(interaction, "🎧 Added to the queue", [f"**{playing_element_name}**"])
 
         if not voice_client.is_playing():
             self.play_queue(interaction, None)
@@ -303,21 +348,38 @@ class QueueCog(Base):
             interaction: The interaction that started the command.
         """
 
-        if await self.is_not_in_channel(interaction):
+        voice_client = await self.get_voice_client(interaction)
+        if voice_client is None:
             return
+
+        if not voice_client.is_playing():
+            await self.send_error(interaction, ["No song currently playing!"])
+            return
+
+        self.skip_next_autoplay = True
+        voice_client.stop()
+        self.now_playing = None
+
+        await self.send_answer(interaction, "🛑 Song stopped")
+
+    @app_commands.command(description="Pause the current song")
+    async def pause(self, interaction: Interaction) -> None:
+        """Pause the song that is currently playing.
+
+        Args:
+            interaction: The interaction that started the command.
+        """
 
         voice_client = await self.get_voice_client(interaction)
         if voice_client is None:
             return
 
         if not voice_client.is_playing():
-            await self.send_embed(interaction, "No song currently playing!")
+            await self.send_error(interaction, ["No song currently playing!"])
             return
 
-        self.skip_next_autoplay = True
-        voice_client.stop()
-
-        await self.send_embed(interaction, "Song stopped")
+        voice_client.pause()
+        await self.send_answer(interaction, "⏸️ Song paused")
 
     @app_commands.command(description="Skip the current song")
     async def skip(self, interaction: Interaction) -> None:
@@ -327,21 +389,18 @@ class QueueCog(Base):
             interaction: The interaction that started the command.
         """
 
-        if await self.is_not_in_channel(interaction):
-            return
-
         voice_client = await self.get_voice_client(interaction)
         if voice_client is None:
             return
 
         if not voice_client.is_playing():
-            await self.send_embed(interaction, "No song currently playing!")
+            await self.send_error(interaction, ["No song currently playing!"])
             return
 
         voice_client.stop()
-        await self.send_embed(interaction, "Song skipped")
+        await self.send_answer(interaction, "⏭️ Song skipped")
 
-    @app_commands.command(description="Resume the queue playback")
+    @app_commands.command(description="Resume the playback")
     async def resume(self, interaction: Interaction) -> None:
         """Resume the playback of the song and if there is no one playing play the next one in the queue.
 
@@ -349,15 +408,21 @@ class QueueCog(Base):
             interaction: The interaction that started the command.
         """
 
-        if await self.is_not_in_channel(interaction):
+        voice_client = await self.get_voice_client(interaction)
+        if voice_client is None:
+            return
+
+        if voice_client.is_paused():
+            voice_client.resume()
+            await self.send_answer(interaction, "▶️ Resuming the song")
             return
 
         if self.queue.length(interaction) == 0:
-            await self.send_embed(interaction, "The queue is empty")
+            await self.send_error(interaction, ["The queue is empty"])
             return
 
         self.play_queue(interaction, None)
-        await self.send_embed(interaction, "Resuming the playback")
+        await self.send_answer(interaction, "▶️ Resuming the playback")
 
     @app_commands.command(name="queue", description="See the current queue")
     # Name changed to avoid collisions with the property `queue`
@@ -369,10 +434,21 @@ class QueueCog(Base):
         """
 
         content = []
-        for song in self.queue.get(interaction):
-            content.append(f"- **{song.title}**")
+        if self.now_playing is not None:
+            content.append(f"Now playing: **{self.now_playing.title}**")
+            content.append("")
 
-        await self.send_embed(interaction, f"Queue ({self.queue.length(interaction)} remaining)", content)
+        length = self.queue.length(interaction)
+
+        if length > 0:
+            content.append("Next:")
+            for song in self.queue.get(interaction):
+                content.append(f"- **{song.title}**")
+
+        if length == 0:
+            content.append("_Queue empty_")
+
+        await self.send_answer(interaction, f"🎹 Queue ({length} songs remaining)", content)
 
     @app_commands.command(description="Adjust the volume")
     async def volume(self, interaction: Interaction, volume: int) -> None:
@@ -383,9 +459,18 @@ class QueueCog(Base):
             volume: The new volume level.
         """
 
-        if volume < 0 or volume > 100:
-            await self.send_embed(interaction, "Volume must be between 0 and 100")
+        voice_client = await self.get_voice_client(interaction)
+        if voice_client is None:
             return
 
-        self.config.volume = volume
-        await self.send_embed(interaction, f"Volume set to {volume}%")
+        if volume < 0:
+            await self.send_error(interaction, ["The requested volume must be at least 0%"])
+            return
+
+        if voice_client.source is None:
+            await self.send_error(interaction, ["The voice client source is not available"])
+            return
+
+        # Every source has a volume handler attach to it so suppressing the mypy error is safe
+        voice_client.source.volume = volume / 100  # type: ignore[attr-defined]
+        await self.send_answer(interaction, f"🔊 Volume level set to {volume}%")